refactor: 응답 코드 사용 및 detail 페이지 타입 상수로 선언
